Don't link directly to the GitHub edit page.
Doing so automatically forks the repository for people that aren't
in the lab, which is irritating if they change their mind and decide
not to edit - and for people not logged in to GitHub, they get an
unhelpful GitHub login page. Link to the GitHub source view instead;
it is easy enough to then click the edit link there.